+++ api/functions/src/send-weekly-email.js
@@ -1,0 +1,8 @@
+const mail = require('../util/email')
+
+module.exports = (ctx) => {
+  mail.sendToList({
+    subject: 'Your Weekly Synonyms',
+    text: 'Congrats on using Simply Synonyms this week!'
+  }, 'weekly-stats')
+}